pyproject.toml
[tool.poetry]
name = "automacao-mandeumzap"
version = "0.1.0"
description = ""
authors = ["lucasrluz <lucasr.luzbr@gmail.com>"]
readme = "README.md"
packages = [{include = "automacao_mandeumzap"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
playwright = "^1.29.1"
python-dotenv = "^0.21.0"
pyinstaller = "^5.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
